Moved points classes
