+++ app/src/main/java/com/github/bgabriel998/softwaredevproject/map/DownloadTopographyTask.java
@@ -5,6 +5,8 @@ import android.util.Log;
 
 import androidx.core.util.Pair;
 
+import com.github.bgabriel998.softwaredevproject.points.UserPoint;
+
 import org.apache.http.HttpResponse;
 import org.apache.http.client.HttpClient;
 import org.apache.http.client.methods.HttpGet;
