+++ app/src/main/java/com/github/bgabriel998/softwaredevproject/points/Point.java
@@ -1,4 +1,4 @@
-package com.github.bgabriel998.softwaredevproject.map;
+package com.github.bgabriel998.softwaredevproject.points;
 
 import androidx.core.util.Pair;
